+++ package/js/views/components/Menu.js
@@ -8,6 +8,12 @@ module.exports = [
         active: true
     },
     {
+        icon: "fa fa-edit fa-fw fa-lg",
+        name: "Editor",
+        href: "/editor",
+        active: false
+    },
+    {
         icon: "fa fa-key fa-fw fa-lg",
         name: "Login",
         href: "/login",
